requirements.txt
PyMuPDF
pandas
SQLAlchemy
psycopg2-binary
fitz
numpy
python-dotenv
redis